Writing records works.
It seems that even though the offsets don't exactly match those in
the original fuelLogDB.pdb file (there are two extra nulls for no
reason I can determine in the original,) pilot-file doesn't seem to
care, as the test_load_and_dump shows its output to be identical
between the original and my loaded-and-dumped recreation.

This means that writing seems to work okay.

I renamed PDB::Resource's .id and PDB::Record's .id to .r_id, because
it's more clear to me.

There's a little design issue that became apparent in the process of
making this output work: the index of record/resources is separate from
the data of the record/resource, but given that metadata is stored with
the index, this internal representation doesn't lend itself to an easy
API.  So I'm making a PDB::Data structure that can be subclassed by
libraries to deal with app-specific data formats, which will provide
access to the metadata and data in a more straightforward way.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>