@records in PalmPDB are instances of either high-level class or PDB::Data
This required some changes in each, and dump() wasn't working with
recent changes in API.  It appears to, now.

Similarly, it's no longer possible to just do PalmPDB.new().  The new
code makes the assumption that it's PDB::AppName.new(), so I changed
the logic in tc_pdbheader.

Maybe this isn't ideal - it'd be nice to be able to open a PDB file
without having to know what app created it.  Even better if it could
figure out what the app is, load the appropriate library, and make a
PDB::AppName class and such out of it.

But that would require changing self partway through parse(), or
starting over after the header's already been read.  Or something.
Right now, I'm not caring much.  I probably will, later.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>