Fixed yaml comparison test.
Really, two yaml structures are equivalent if their to_s's are
equivalent.  And it helps to compare strings with strings, not File
objects.

I also am pretty sure that the yaml file now corresponds to the actual
data I'm trying to test for.  I should double check this against either
the data as the palm (or emulated palm) shows it.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>