Work on supporting adding records.
This commit breaks the rule that all my changes should be simple,
small, and address only one thing at a time.  The fact is, work on
this code is getting harder as the file structure needs to be fixed,
and various code split out into different files.

So this is a half-finished WIP commit.  I won't make a habit of it.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>