Implementing records=
This allows the set of records to be assigned from either an array or
a hash.  In the process, I expanded []= to consider cases where a
record is being overwritten instead of simply added, and delegated
the process of actually adding a record to it, from <<.

It seems to work so far, if tc_modify is any indicator.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>