Change Kernel.const_get_from_string to String.get_full_const
This simplifies a lot of the code that needs to use this method,
partly because I move the responsibility for rescuing on NameError
from the users' code to the common method instead.

On seeing a similar method in Merb, I realized how much of an
improvement this is on my previous technique.

Signed-off-by: Chris Riddoch <chris@syntacticsugar.org>